Fixed "FROM" formatting in SQL panel layout.
We always start a new line for "FROM" now. Previous code was assuming MySQL
quoting of identifiers (MySQL is a bit non-standard in using backquotes) and
was assuming the last thing before the FROM would be a quoted identifier, which
wasn't always true, particularly when extra() is used on querysets.

Signed-off-by: Rob Hudson <rob@cogit8.org>