Change how JS libs are loaded
This change makes Django-Debug-Toolbar use the existing version of jQuery
on the page if one exists, and only loads jQuery if it isn't already loaded.
It also explicitly loads the jQuery cookie plugin and the toolbar script
as separate files. Finally, we are no longer compressing/minimizing scripts;
this is for debug purposes, and presumably will not be served in production.