Don't ignore templates without an origin
We're running Django 1.7 with django-jinja. For whatever reasons, the template origin isn't set, and the panel shows no template context whatsoever. Upon investigation, I discovered that the template rendered signal is indeed fired, but the template is skipped as no origin is set.

I deny any understanding of the existing code. I do not know why we skip templates without an origin. But without this change, we aren't getting a template context in the panel, and with it, we are. So at first sight, it looks like it's safe to do. But somebody more qualified than me has to be the judge on that.

Furthermore, this code marks the 'No origin' string as translatable.